backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-cors==5.0.0
flask-jwt-extended==4.7.1
pymysql==1.1.1
python-dotenv==1.0.1
cryptography==44.0.0